LAGOSUS Identifier links

Includes single or multiple identifiers per lake from other commonly used national-scale data products including: Water Quality Portal (WQP), Geographic Names Information System (GNIS), NHD MR Plusv2, LAGOS-NE, and the EPA National Lake Assessment surveys from 2007 and 2012.

Details

Variable name	Variable description	Units

lagoslakeid	unique lake identifier developed by LAGOS-US

lake_nhdid	the unique Permanent_Identifier from the NHD for each LAGOS lake

lake_reachcode	reach code assigned to the water feature by the NHD

lake_namegnis	lake name from the GNIS database

lake_namelagos	lake name from a combination of data sources; examples are GNIS, WQP, etc.

lake_county	name of county containing the focal lake polygon centroid

lake_countyfips	Federal Information Processing Standard, FIPS, code for the county containing the focal lake polygon centroid

lake_lat_decdeg	latitude of central point inside the NHD lake polygon in decimal degrees; NAD83 projection	decimal degree

lake_lon_decdeg	longitude of central point inside the NHD lake polygon in decimal degrees; NAD83 projection	decimal degree

lake_centroidstate	two-letter postal abbreviation of the state containing the lake central point

nhdhr_area_sqkm	area of areal feature based on Albers Equal Area, as assigned in NHD	square kilometer

nhdhr_fdate	date of last feature modification in National Hydrography Dataset

nhdhr_gnisid	unique identifier assigned by GNIS

lagosus_legacysiteid	unique site or lake identifier from the original LAGOS-US water quality source dataset--this is not standardized and each source dataset has its own system

lagosus_legacysitelabel	column name or field name used by the source dataset to label the legacy lake identifier

lagosus_legacyprogram	program name assigned in LAGOS-US to identify the source dataset provider

wqp_monitoringlocationidentifier	a designator used to describe the unique name, number, or code assigned to identify the monitoring location

wqp_monitoringlocationname	the designator specified by the sampling organization for the site at which sampling or other activities are conducted

wqp_providername	source system that provided data to the Water Quality Portal; examples are NWIS, STORET, STEWARDS, etc.

nhdplusv2_comid	common identifier of the NHD Waterbody feature

nhdplusv2_reachcode	NHDplus v2 reach code assigned to the water feature.

nhdplusv2_area_sqkm	feature area in square kilometers, as assigned by NHD	square kilometer

lagosne_lagoslakeid	unique lake identifier developed for LAGOS-NE

lagosne_legacysiteid	unique site/lake identifier from the original LAGOS-NE water quality source dataset--this is not standardized and each source dataset has its own system

nla2007_siteid	lake site identifier from National Lakes Assessment 2007 survey

nla2012_siteid	lake site identifier from National Lakes Assessment 2012 survey

nhdplusv2_lakes_n	count of NHDPlusV2 polygons matching the LAGOS-US lagoslakeid	count

lagosne_lakes_n	count of LAGOS-NE lagoslakeids matching the LAGOS-US lagoslakeid	count

wqp_sites_n	count of WQP sites within the LAGOS-US lake	count

lagosus_legacyids_n	count of all LAGOS-US legacy ids linked to the lagoslakeid in this table; these ids include, for example, original sampling site identifiers	count