LAGOSUS Lake watersheds

Includes identifiers, location, and geometry for the calculated watersheds (WS) and network watersheds (NWS).

Details

Variable name	Variable description	Units

lagoslakeid	unique lake identifier developed by LAGOS-US	NULL

ws_zoneid	unique WS watershed identifier assigned by LAGOS-US; identifier is equivalent to the focal lake lagoslakeid	NULL

nws_zoneid	unique NWS watershed identifier assigned by LAGOS-US; identifier is equivalent to the focal lake lagoslakeid	NULL

ws_subtype	watershed subtype class for the watershed; LC equals local catchment, DWS equals drainage-watershed, IDWS equals inter-drainage-lake watershed	NULL

ws_equalsnws	flag indicating that the ws is equivalent to the nws	NULL

ws_onlandborder	flag indicating zone is adjacent to or crosses the border with Canada or Mexico	NULL

ws_oncoast	flag indicating the zone is adjacent to the coastline	NULL

ws_inusa_pct	percent of the zone polygon that is within the US border	percent

ws_includeshu4inlet	flag indicating the WS extends beyond the focal HU4 and is connected via an inlet to an adjoining HU4	NULL

ws_ismultipart	flag indicating that the focal zone polygon is multipart	NULL

ws_sliverflag	flag indicating extreme cases where the watershed is comprised of slivers due to flat or constrained terrain that may not accurately reflect surface drainage	NULL

ws_sinkerrorflag	flag indicating an error that applies to a small subset of watersheds that erroneously include in the WS polygon an upstream lake with only in inlet that is greater than or equal to 10 ha, along with its respective WS	NULL

ws_divergenceerrorflag	flag indicating an error that applies to a small subset of watersheds when the presence of a second lake greater than or equal to 10ha located below a common network divergence led to the erroneous erasure of an overlapping watershed area upstream of that divergence from the focal lake watershed	NULL

nws_divergenceerrorflag	flag indicating an error that applies to a small subset of watersheds when the presence of a second lake greater than or equal to 10ha located below a common network divergence led to the erroneous erasure of an overlapping watershed area upstream of that divergence from the focal lake network watershed	NULL

nws_onlandborder	flag indicating zone is adjacent to or crosses the border with Canada or Mexico	NULL

nws_oncoast	flag indicating the zone is adjacent to the coastline	NULL

nws_inusa_pct	percent of the zone polygon that is within the US border	percent

nws_includeshu4inlet	flag indicating the NWS extends beyond the focal HU4 and is connected via an inlet to an adjoining HU4	NULL

nws_ismultipart	flag indicating that the focal zone polygon is multipart	NULL

ws_states	two-letter postal abbreviations of all states intersecting the zone polygon	NULL

ws_focallakewaterarea_ha	surface area of focal lake for which the WS is calculated	hectare

ws_area_ha	area of zone polygon	hectare

ws_perimeter_m	perimeter of zone polygon including perimeter of any internal holes	meter

ws_lake_arearatio	ratio between WS area and lake water area	NULL

ws_mbgconhull_length_m	the maximum possible distance between any two vertices of the minimum bounding geometry convex hull that encloses the WS polygon	meter

ws_mbgconhull_width_m	the maximum width of the WS convex hull polygon along an axis perpendicular to the axis defined by the convex hull length	meter

ws_mbgconhull_orientation_deg	orientation of the line defining the convex hull length; calculated as decimal degrees clockwise from north	degree

ws_meanwidth_m	mean width of the WS polygon; calculated as ws_area_ha converted to hectares and divided by ws_mbgconhull_length_m	meter

ws_lat_decdeg	latitude of the centroid of the focal lake polygon in decimal degrees; NAD83 projection	decimaldegree

ws_lon_decdeg	longitude of the centroid of the focal lake polygon in decimal degrees; NAD83 projection	decimaldegree

nws_states	two-letter postal abbreviations of all states intersecting the zone polygon	NULL

nws_focallakewaterarea_ha	surface area of focal lake for which the NWS is calculated	hectare

nws_area_ha	area of zone polygon	hectare

nws_perimeter_m	perimeter of zone polygon including perimeters of any internal holes	meter

nws_lake_arearatio	ratio between NWS area and lake water area	NULL

nws_mbgconhull_length_m	the maximum possible distance between any two vertices of the minimum bounding geometry convex hull that encloses the NWS polygon	meter

nws_mbgconhull_width_m	the maximum width of the NWS convex hull polygon along an axis perpendicular to the axis defined by the convex hull length	meter

nws_mbgconhull_orientation_deg	orientation of the line defining the convex hull length; calculated as decimal degrees clockwise from north	degree

nws_meanwidth_m	mean width of the NWS polygon; calculated as nws_area_ha converted to hectares and divided by nws_mbgconhull_length_m	meter

nws_lat_decdeg	latitude of the centroid of the focal lake polygon in decimal degrees; NAD83 projection	decimal degree

nws_lon_decdeg	longitude of the centroid of the focal lake polygon in decimal degrees; NAD83 projection	decimal degree