LAGOSUS Lake characteristics

Includes derived metrics for each lake such as lake geometry (e.g., area, perimeter, shape metrics), lake connectivity (e.g., lake connectivity class, number of upstream lakes), and whether the lake is located within an area that was glaciated during the Late Wisconsin glaciation.

Details

Variable name	Variable description	Units

lagoslakeid	unique lake identifier developed by LAGOS-US

lake_waterarea_ha	surface area of lake waterbody polygon from NHD (excludes islands)	hectare

lake_totalarea_ha	surface area within outer boundary of lake waterbody polygon from NHD (includes islands)	hectare

lake_islandarea_ha	surface area of islands within outer boundary of lake waterbody polygon from NHD	hectare

lake_perimeter_m	perimeter of outer boundary of lake waterbody polygon from the NHD (excludes islands)	meter

lake_islandperimeter_m	perimeter of islands within the lake waterbody polygon from the NHD	meter

lake_shorelinedevfactor	shoreline development factor = lake_perimeter / (2*?(?*waterarea))

lake_mbgconhull_length_m	the longest distance entirely within the lake polygon between any two vertices of the minimum bounding geometry convex hull that encloses the lake polygon	meter

lake_mbgconhull_width_m	the shortest distance entirely within the lake polygon between any two vertices of the minimum bounding geometry convex hull that encloses the lake polygon	meter

lake_mbgconhull_orientation_deg	orientation (from 0 to 180 decimal degrees clockwise from north) of the longest line connecting two vertices entirely within both the convex hull and the lake polygon	degree

lake_mbgrect_length_m	length of the minimum bounding geometry rectangle by area enclosing the lake polygon	meter

lake_mbgrect_width_m	width of the minimum bounding geometry rectangle by area enclosing the lake polygon	meter

lake_mbgrect_arearatio	ratio between lake_waterarea_ha (* 10000 m^2/ha) and the area of the minimum bounding rectangle by area of the lake polygon; value near 1 indicates lake shape is rectangular

lake_meanwidth_m	mean width of the lake calculated as lake water area (lake_area_ha* 10000 m^2/ha) divided by the convex hull length of the focal lake polygon (lake_mbgconhull_length_m)	meter

lake_connectivity_class	hydrologic connectivity class of the focal lake determined from the NHD network considering both permanent and intermittent/ephemeral flow

lake_connectivity_fluctuates	indicates whether the lake's connectivity classification depends on non-permanent flow

lake_connectivity_permanent	hydrologic connectivity class of the focal lake determined from the NHD network considering only permanent flow

lake_lakes1ha_upstream_ha	total area of lakes >= 1 ha upstream of the focal lake, connected via surface streams	hectare

lake_lakes4ha_upstream_ha	total area of lakes >= 4 ha upstream of the focal lake, connected via surface streams	hectare

lake_lakes10ha_upstream_ha	total area of lakes >= 10 ha upstream of the focal lake, connected via surface streams	hectare

lake_lakes1ha_upstream_n	count of lakes >= 1 ha upstream of the focal lake, connected via surface streams	count

lake_lakes4ha_upstream_n	count of lakes >= 4 ha upstream of the focal lake, connected via surface streams	count

lake_lakes10ha_upstream_n	count of lakes >= 10 ha upstream of the focal lake, connected via surface streams	count

lake_glaciatedlatewisc	glaciation status during the Late Wisconsin glaciation (status = Glaciated if any part of the lake polygon is inside the glacial extent)