LAGOSUS Lake information

LAGOSUS Lake information

Details

lagoslakeid	LAGOS sourced unique identifier for the focal lake

lake_nhdid	the unique 'Permanent_identifier' from the NHD for each LAGOS lake

lake_nhdfcode	NHD five-digit integer code with feature code plus additional characteristics/values

lake_nhdftype	NHD three-digit integer code providing a unique identifier of feature type for the waterbody

lake_reachcode	NHD reachcode for stream reach intersecting the lake from NHD stream network

lake_namegnis	lake name from the gnis database

lake_namelagos	lake name from a combination of data sources (gnis, wqp, etc.)

♦	flag indicating zone is adjacent to or crosses the border with Canada or Mexico

lake_ismultipart	flag indicating that the focal lake polygon is multipart

lake_missingws	flag indicating that the lake's watershed (ws or nws) was not delineated

lake_shapeflag	flag indicating lake polygon shape is excessively angular (e.g., triangle, rectangle) or elongate (very thin relative to length)

lake_lat_decdeg	latitude of centroid of the NHD lake polygon in decimal degrees (NAD83)

lake_lon_decdeg	longitude of centroid of the NHD lake polygon in decimal degrees (NAD83)

lake_centroidstate	abbreviation of state containing the lake centroid

lake_states	abbreviation(s) of state(s) intersecting the zone polygon

lake_huc12	HUC12 of the WBD HU12 watershed containing the lake

buff100_zoneid	unique 100m buffer zone identifier equivalent to the focal lake lagoslakeid assigned by LAGOS

buff500_zoneid	unique 500m buffer zone identifier equivalent to the focal lake lagoslakeid assigned by LAGOS

ws_zoneid	unique zone identifier equivalent to the focal lake lagoslakeid assigned by LAGOS

nws_zoneid	unique zone identifier equivalent to the focal lake lagoslakeid assigned by LAGOS

hu12_zoneid	unique HU12 zone identifier assigned by LAGOS for zone containing the focal lake

hu8_zoneid	unique HU8 zone identifier assigned by LAGOS for zone containing the focal lake

hu4_zoneid	unique HU4 zone identifier assigned by LAGOS for zone containing the focal lake

county_zoneid	unique county zone identifier assigned by LAGOS for zone containing the focal lake

state_zoneid	unique state zone identifier assigned by LAGOS for zone containing the focal lake

epanutr_zoneid	unique EPA Nutrient Ecoregion zone identifier assigned by LAGOS for zone containing the focal lake

omernik3_zoneid	unique Omernik Level III Ecoregion zone identifier assigned by LAGOS for zone containing the focal lake

wwf_zoneid	unique WWF zone identifier assigned by LAGOS for zone containing the focal lake

mlra_zoneid	unique Major Land Resource Area zone identifier assigned by LAGOS for zone containing the focal lake

bailey_zoneid	unique Bailey Ecoregion zone identifier assigned by LAGOS for zone containing the focal lake

neon_zoneid	unique NEON zone identifier assigned by LAGOS for zone containing the focal lake