LAGOSUS Data sources

LAGOSUS Data sources

Variable name	Variable description	Units

source_code	a short codename used for the source dataset in LAGOS-US variables or metadata

source_name	the name of the source dataset

source_shortname	the abbreviated name used to refer to the source dataset in LAGOS-US documentation

source_description	a brief description of the source dataset

source_provider	the provider of the source dataset (government agency, publication citation, etc.)

source_citation	the citation for the source dataset, including the download link used to acquire the dataset at the time of LAGOS-US creation

source_url	The principal website associated with the source dataset. This may not be the same as the download URL included in the citation.

source_metadata_url	a website URL where metadata and data use guidance for the source dataset can be found

source_spatial_resolution	the spatial resolution of the source dataset, if applicable

source_time_period	the time period captured by the source dataset, if applicable

source_access_date	the date(s) the source dataset was acquired for use in creating LAGOS-US

source_data_type	the data value type contained in the source dataset, if applicable