LAGOSUS Lake watersheds

lagoslakeid	LAGOS sourced unique identifier for the focal lake of the zone

ws_zoneid	unique zone identifier equivalent to the focal lake lagoslakeid assigned by LAGOS

nws_zoneid	unique zone identifier equivalent to the focal lake lagoslakeid assigned by LAGOS

ws_subtype	LAGOS watershed sub-type classification for the watershed

ws_equalsnws	flag indicating that the focal zone polygon is multipart

ws_onlandborder	flag indicating zone is adjacent to or crosses the border with Canada or Mexico

ws_oncoast	flag indicating the zone is adjacent to the coastline

ws_inusa_pct	percent of the zone (i.e., polygon) that is within the US border

ws_includeshu4inlet	flag indicating the ws extends beyond the focal HU4 and is connected via an inlet to an adjoining HU4

ws_ismultipart	flag indicating that the zonal polygon is multipart

ws_sliver_flag	NA

nws_onlandborder	flag indicating zone is adjacent to or crosses the border with Canada or Mexico

nws_oncoast	flag indicating the zone is adjacent to the coastline

nws_inusa_pct	percent of the zone (i.e., polygon) that is within the US border

nws_includeshu4inlet	flag indicating the nws extends beyond the focal HU4 and is connected via an inlet to an adjoining HU4

nws_ismultipart	flag indicating that the zonal polygon is multipart

ws_states	abbreviation(s) of state(s) intersecting the zone polygon

ws_focallakewaterarea_ha	surface area of focal lake for which the watershed (WS) is calculated

ws_area_ha	area of focal zone polygon

ws_perimeter_m	perimeter of the focal zone polygon including perimeters of any internal 'holes'

ws_lake_arearatio	NA

ws_mbgconhull_length_m	the longest distance between any two vertices of the minimum bounding geometry convex hull enclosing the zonal polygon

ws_mbgconhull_width_m	the shortest distance between any two vertices of the minimum bounding geometry convex hull enclosing the zonal polygon

ws_mbgconhull_orientation_deg	orientation (from 0 to 180 degrees) of the line connecting the antipodal pairs defining the longest distance within the minimum bounding geometry convex hull enclosing the zonal polygon

ws_meanwidth_m	ws area (* 10000 m^2/ha) divided by the ws convex hull length of the focal zone polygon

nws_states	abbreviation(s) of state(s) intersecting the zone polygon

nws_focallakewaterarea_ha	surface area of focal lake for which the watershed (NWS) is calculated

nws_area_ha	area of zone polygon

nws_perimeter_m	perimeter of the focal zone polygon including perimeters of any internal 'holes'

nws_lake_arearatio	NA

nws_mbgconhull_length_m	the longest distance between any two vertices of the minimum bounding geometry convex hull enclosing the zonal polygon

nws_mbgconhull_width_m	the shortest distance between any two vertices of the minimum bounding geometry convex hull enclosing the zonal polygon

nws_mbgconhull_orientation_deg	orientation (from 0 to 180 degrees) of the line connecting the antipodal pairs defining the longest distance within the minimum bounding geometry convex hull enclosing the zonal polygon

nws_meanwidth_m	nws area (* 10000 m^2/ha) divided by the nws convex hull length of the focal zone polygon