LAGOSUS Chemical and physical data

LAGOSUS Chemical and physical data

Variable name	Variable description	Units

lagoslakeid	unique lake identifier developed by LAGOSUS	NULL

eventid_epi	unique event identifier that is a combination of programid, lakeid, and date for each sampling event in LAGOS	NULL

site_legacyid	unique sample site identifier provided by the source of the data; this is not standardardized and each source dataset has its own system	NULL

program_id	unique integer program identifier assigned by LAGOSUS	NULL

program_name	name of sampling program that collected the data	NULL

event_date	date, as YYYY-MM-DD, when the lake was sampled by the source program	YYYY-MM-DD

site_id	unique sample site identifier created by the LAGOSUS	NULL

site_lat_decdeg	latitude of the sample site	Decimal Degrees

site_lon_decdeg	longitude of the sample site	Decimal Degrees

program_type	type of program that provided the data	NULL

alk_ueql	alkalinity, microequivalents per liter	Microequivalents per Liter

alk_qualifier_legacyflag	alkalinity data flags from the source program	NULL

alk_qualifier_legacyflagdesc	alkalinity data flag descriptions from the source program	NULL

alk_censorcode	the LAGOSUS assigned censor code for alkalinity that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself	NULL

alk_detectionlimit_legacyvalue	alkalinity detection limit used by the source program	NULL

alk_detectionlimit_legacyunit	alkalinity detection limit unit used by the source program	NULL

alk_method	alkalinity analytical procedure, from a standards body if available	NULL

ca_mgl	calcium, dissolved, milligrams per liter	Milligrams per Liter

ca_qualifier_legacyflag	calcium, dissolved data flags from the source program	NULL

ca_qualifier_legacyflagdesc	calcium, dissolved data flag descriptions from the source program	NULL

ca_censorcode	the LAGOSUS assigned censor code for calcium, dissolved that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself	NULL

ca_detectionlimit_legacyvalue	calcium, dissolved detection limit used by the source program	NULL

ca_detectionlimit_legacyunit	calcium, dissolved detection limit unit used by the source program	NULL

ca_method	calcium, dissolved analytical procedure, from a standards body if available	NULL

cl_mgl	chloride, dissolved, milligrams per liter	Milligrams per Liter

cl_qualifier_legacyflag	chloride, dissolved data flags from the source program	NULL

cl_qualifier_legacyflagdesc	chloride, dissolved data flag descriptions from the source program	NULL

cl_censorcode	the LAGOSUS assigned censor code for chloride, dissolved that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself	NULL

cl_detectionlimit_legacyvalue	chloride, dissolved detection limit used by the source program	NULL

cl_detectionlimit_legacyunit	chloride, dissolved detection limit unit used by the source program	NULL

cl_method	chloride, dissolved analytical procedure, from a standards body if available	NULL

spcond_uscm	specific conductance, microsiemens per cm at 25C	Microsiemens per Centimeter

spcond_qualifier_legacyflag	specific conductance data flags from the source program	NULL

spcond_qualifier_legacyflagdesc	specific conductance data flag descriptions from the source program	NULL

spcond_censorcode	the LAGOSUS assigned censor code for specific conductance that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself	NULL

spcond_detectionlimit_legacyvalue	specific conductance detection limit used by the source program	NULL

spcond_detectionlimit_legacyunit	specific conductance detection limit unit used by the source program	NULL

spcond_method	specific conductance analytical procedure, from a standards body if available	NULL

do_mgl	oxygen, dissolved, milligrams per liter	Milligrams per Liter

do_qualifier_legacyflag	oxygen, dissolved data flags from the source program	NULL

do_qualifier_legacyflagdesc	oxygen, dissolved data flag descriptions from the source program	NULL

do_censorcode	the LAGOSUS assigned censor code for dissolved oxygen that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself	NULL

do_detectionlimit_legacyvalue	oxygen, dissolved detection limit used by the source program	NULL

do_detectionlimit_legacyunit	oxygen, dissolved detection limit unit used by the source program	NULL

do_method	oxygen, dissolved analytical procedure, from a standards body if available	NULL

mg_mgl	magnesium, dissolved, milligrams per liter	Milligrams per Liter

mg_qualifier_legacyflag	magnesium, dissolved data flags from the source program	NULL

mg_qualifier_legacyflagdesc	magnesium, dissolved data flag descriptions from the source program	NULL

mg_censorcode	the LAGOSUS assigned censor code for magnesium, dissolved that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself	NULL

mg_detectionlimit_legacyvalue	magnesium, dissolved detection limit used by the source program	NULL

mg_detectionlimit_legacyunit	magnesium, dissolved detection limit unit used by the source program	NULL

mg_method	magnesium, dissolved analytical procedure, from a standards body if available	NULL

ph_eq	pH, equilibrated	NULL

ph_eq_qualifier_legacyflag	pH, equilibrated data flags from the source program	NULL

ph_eq_qualifier_legacyflagdesc	pH, equilibrated data flag descriptions from the source program	NULL

ph_eq_censorcode	the LAGOSUS assigned censor code for ph_eq that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself	NULL

ph_eq_detectionlimit_legacyvalue	pH, equilibrated detection limit used by the source program	NULL

ph_eq_detectionlimit_legacyunit	pH, equilibrated detection limit unit used by the source program	NULL

ph_eq_method	pH, equilibrated analytical procedure, from a standards body if available	NULL

ph_field	pH, field or closed	NULL

ph_field_qualifier_legacyflag	pH, field or closed data flags from the source program	NULL

ph_field_qualifier_legacyflagdesc	pH, field or closed data flag descriptions from the source program	NULL

ph_field_censorcode	the LAGOSUS assigned censor code for ph_field that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself	NULL

ph_field_detectionlimit_legacyvalue	pH, field or closed detection limit used by the source program	NULL

ph_field_detectionlimit_legacyunit	pH, field or closed detection limit unit used by the source program	NULL

ph_field_method	pH, field or closed analytical procedure, from a standards body if available	NULL

salinity_mgl	salinity, milligrams per liter	Milligrams per Liter

salinity_qualifier_legacyflag	salinity data flags from the source program	NULL

salinity_qualifier_legacyflagdesc	salinity data flag descriptions from the source program	NULL

salinity_censorcode	the LAGOSUS assigned censor code for salinity that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself	NULL

salinity_detectionlimit_legacyvalue	salinity detection limit used by the source program	NULL

salinity_detectionlimit_legacyunit	salinity detection limit unit used by the source program	NULL

salinity_method	salinity analytical procedure, from a standards body if available	NULL

so4_mgl	sulfate, dissolved, milligrams per liter as SO4	Milligrams per Liter

so4_qualifier_legacyflag	sulfate, dissolved data flags from the source program	NULL

so4_qualifier_legacyflagdesc	sulfate, dissolved data flag descriptions from the source program	NULL

so4_censorcode	the LAGOSUS assigned censor code for sulfate, dissolved that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself	NULL

so4_detectionlimit_legacyvalue	sulfate, dissolved detection limit used by the source program	NULL

so4_detectionlimit_legacyunit	sulfate, dissolved detection limit unit used by the source program	NULL

so4_method	sulfate, dissolved analytical procedure, from a standards body if available	NULL

temp_degc	temperature, degrees Celsius	Degrees Celcius

temp_qualifier_legacyflag	temperature data flags from the source program	NULL

temp_qualifier_legacyflagdesc	temperature data flag descriptions from the source program	NULL

temp_censorcode	the LAGOSUS assigned censor code for temperature that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself	NULL

temp_detectionlimit_legacyvalue	temperature detection limit used by the source program	NULL

temp_detectionlimit_legacyunit	temperature detection limit unit used by the source program	NULL

temp_method	temperature analytical procedure, from a standards body if available	NULL