LAGOSUS Contaminant data

LAGOSUS Contaminant data

Variable name	Variable description	Units

eventid_epi	unique event identifier that is a combination of programid, lakeid, and date for each sampling event in LAGOS	NULL

site_legacyid	unique sample site identifier provided by the source of the data; this is not standardardized and each source dataset has its own system	NULL

program_id	unique integer program identifier assigned by LAGOSUS	NULL

program_name	name of sampling program that collected the data	NULL

event_date	date, as YYYY-MM-DD, when the lake was sampled by the source program	YYYY-MM-DD

site_id	unique sample site identifier created by the LAGOSUS	NULL

lagoslakeid	unique lake identifier developed by LAGOSUS	NULL

site_lat_decdeg	latitude of the sample site	Decimal Degrees

site_lon_decdeg	longitude of the sample site	Decimal Degrees

program_type	type of program that provided the data	NULL

al_tot_ugl	aluminum, total, micrograms per liter	Micrograms per Liter

al_tot_qualifier_legacyflag	aluminum, total data flags from the source program	NULL

al_tot_qualifier_legacyflagdesc	aluminum, total data flag descriptions from the source program	NULL

al_tot_censorcode	the LAGOSUS assigned censor code for aluminum, total that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself	NULL

al_tot_detectionlimit_legacyvalue	aluminum, total detection limit used by the source program	NULL

al_tot_detectionlimit_legacyunit	aluminum, total detection limit unit used by the source program	NULL

al_tot_method	aluminum, total analytical procedure, from a standards body if available	NULL

al_diss_ugl	aluminum, dissolved, micrograms per liter	Micrograms per Liter

al_diss_qualifier_legacyflag	aluminum, dissolved data flags from the source program	NULL

al_diss_qualifier_legacyflagdesc	aluminum, dissolved data flag descriptions from the source program	NULL

al_diss_censorcode	the LAGOSUS assigned censor code for aluminum, dissolved that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself	NULL

al_diss_detectionlimit_legacyvalue	aluminum, dissolved detection limit used by the source program	NULL

al_diss_detectionlimit_legacyunit	aluminum, dissolved detection limit unit used by the source program	NULL

al_diss_method	aluminum, dissolved analytical procedure, from a standards body if available	NULL

as_tot_ugl	arsenic, total, micrograms per liter	Micrograms per Liter

as_tot_qualifier_legacyflag	arsenic, total data flags from the source program	NULL

as_tot_qualifier_legacyflagdesc	arsenic, total data flag descriptions from the source program	NULL

as_tot_censorcode	the LAGOSUS assigned censor code for arsenic, total that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself	NULL

as_tot_detectionlimit_legacyvalue	arsenic, total detection limit used by the source program	NULL

as_tot_detectionlimit_legacyunit	arsenic, total detection limit unit used by the source program	NULL

as_tot_method	arsenic, total analytical procedure, from a standards body if available	NULL

as_diss_ugl	arsenic, dissolved, micrograms per liter	Micrograms per Liter

as_diss_qualifier_legacyflag	arsenic, dissolved data flags from the source program	NULL

as_diss_qualifier_legacyflagdesc	arsenic, dissolved data flag descriptions from the source program	NULL

as_diss_censorcode	the LAGOSUS assigned censor code for arsenic, dissolved that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself	NULL

as_diss_detectionlimit_legacyvalue	arsenic, dissolved detection limit used by the source program	NULL

as_diss_detectionlimit_legacyunit	arsenic, dissolved detection limit unit used by the source program	NULL

as_diss_method	arsenic, dissolved analytical procedure, from a standards body if available	NULL

atz_tot_ugl	atrazine, total, micrograms per liter	Micrograms per Liter

atz_tot_qualifier_legacyflag	atrazine, total data flags from the source program	NULL

atz_tot_qualifier_legacyflagdesc	atrazine, total data flag descriptions from the source program	NULL

atz_tot_censorcode	the LAGOSUS assigned censor code for atrazine, total that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself	NULL

atz_tot_detectionlimit_legacyvalue	atrazine, total detection limit used by the source program	NULL

atz_tot_detectionlimit_legacyunit	atrazine, total detection limit unit used by the source program	NULL

atz_tot_method	atrazine, total analytical procedure, from a standards body if available	NULL

atz_diss_ugl	atrazine, dissolved, micrograms per liter	Micrograms per Liter

atz_diss_qualifier_legacyflag	atrazine, dissolved data flags from the source program	NULL

atz_diss_qualifier_legacyflagdesc	atrazine, dissolved data flag descriptions from the source program	NULL

atz_diss_censorcode	the LAGOSUS assigned censor code for atrazine, dissolved that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself	NULL

atz_diss_detectionlimit_legacyvalue	atrazine, dissolved detection limit used by the source program	NULL

atz_diss_detectionlimit_legacyunit	atrazine, dissolved detection limit unit used by the source program	NULL

atz_diss_method	atrazine, dissolved analytical procedure, from a standards body if available	NULL

ecoli_cfu100ml	E. coli CFU, colony forming units per 100ml	Colony Forming Units per 100 Milliliters

ecoli_cfu_qualifier_legacyflag	E. coli CFU data flags from the source program	NULL

ecoli_cfu_qualifier_legacyflagdesc	E. coli CFU data flag descriptions from the source program	NULL

ecoli_cfu_censorcode	the LAGOSUS assigned censor code for E. coli, cfu that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself	NULL

ecoli_cfu_detectionlimit_legacyvalue	E. coli CFU detection limit used by the source program	NULL

ecoli_cfu_detectionlimit_legacyunit	E. coli CFU detection limit unit used by the source program	NULL

ecoli_cfu_method	E. coli CFU analytical procedure, from a standards body if available	NULL

ecoli_mpn100ml	E. coli MPN, most probable number per 100ml	Most Probable Number per 100 Milliliters

ecoli_mpn_qualifier_legacyflag	E. coli MPN data flags from the source program	NULL

ecoli_mpn_qualifier_legacyflagdesc	E. coli MPN data flag descriptions from the source program	NULL

ecoli_mpn_censorcode	the LAGOSUS assigned censor code for E. coli, MPN that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself	NULL

ecoli_mpn_detectionlimit_legacyvalue	E. coli MPN detection limit used by the source program	NULL

ecoli_mpn_detectionlimit_legacyunit	E. coli MPN detection limit unit used by the source program	NULL

ecoli_mpn_method	E. coli MPN analytical procedure, from a standards body if available	NULL

hg_tot_ngl	mercury, total, nanograms per liter	Nanograms per Liter

hg_tot_qualifier_legacyflag	mercury, total data flags from the source program	NULL

hg_tot_qualifier_legacyflagdesc	mercury, total data flag descriptions from the source program	NULL

hg_tot_censorcode	the LAGOSUS assigned censor code for mercury, total that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself	NULL

hg_tot_detectionlimit_legacyvalue	mercury, total detection limit used by the source program	NULL

hg_tot_detectionlimit_legacyunit	mercury, total detection limit unit used by the source program	NULL

hg_tot_method	mercury, total analytical procedure, from a standards body if available	NULL

hg_totdiss_ngl	mercury, total dissolved, nanograms per liter	Nanograms per Liter

hg_totdiss_qualifier_legacyflag	mercury, total dissolved data flags from the source program	NULL

hg_totdiss_qualifier_legacyflagdesc	mercury, total dissolved data flag descriptions from the source program	NULL

hg_totdiss_censorcode	the LAGOSUS assigned censor code for mercury, dissolved that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself	NULL

hg_totdiss_detectionlimit_legacyvalue	mercury, total dissolved detection limit used by the source program	NULL

hg_totdiss_detectionlimit_legacyunit	mercury, total dissolved detection limit unit used by the source program	NULL

hg_totdiss_method	mercury, total dissolved analytical procedure, from a standards body if available	NULL

mehg_tot_ngl	methylmercury, total, nanograms per liter	Nanograms per Liter

mehg_tot_qualifier_legacyflag	methylmercury, total data flags from the source program	NULL

mehg_tot_qualifier_legacyflagdesc	methylmercury, total data flag descriptions from the source program	NULL

mehg_tot_censorcode	the LAGOSUS assigned censor code for methylmercury, total that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself	NULL

mehg_tot_detectionlimit_legacyvalue	methylmercury, total detection limit used by the source program	NULL

mehg_tot_detectionlimit_legacyunit	methylmercury, total detection limit unit used by the source program	NULL

mehg_tot_method	methylmercury, total analytical procedure, from a standards body if available	NULL

se_tot_ugl	selenium, total, micrograms per liter	Micrograms per Liter

se_tot_qualifier_legacyflag	selemium, total data flags from the source program	NULL

se_tot_qualifier_legacyflagdesc	selemium, total data flag descriptions from the source program	NULL

se_tot_censorcode	the LAGOSUS assigned censor code for seleneium, total that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself	NULL

se_tot_detectionlimit_legacyvalue	selenium, total detection limit used by the source program	NULL

se_tot_detectionlimit_legacyunit	selenium, total detection limit unit used by the source program	NULL

se_tot_method	selenium, total analytical procedure, from a standards body if available	NULL

se_diss_ugl	selenium, dissolved, micrograms per liter	Micrograms per Liter

se_diss_qualifier_legacyflag	selenium, dissolved data flags from the source program	NULL

se_diss_qualifier_legacyflagdesc	selenium, dissolved data flag descriptions from the source program	NULL

se_diss_censorcode	the LAGOSUS assigned censor code for selenium, dissolved that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself	NULL

se_diss_detectionlimit_legacyvalue	selenium, dissolved detection limit used by the source program	NULL

se_diss_detectionlimit_legacyunit	selenium, dissolved detection limit unit used by the source program	NULL

se_diss_method	selenium, dissolved analytical procedure, from a standards body if available	NULL