LAGOSUS Lake information

LAGOSUS Lake information

Variable name	Variable description	Units

lagoslakeid	unique lake identifier developed by LAGOS-US

lake_nhdid	the unique 'Permanent_identifier' from the NHD for each LAGOS lake

lake_nhdfcode	NHD five-digit integer code with feature code plus additional characteristics/values

lake_nhdftype	NHD three-digit integer code providing a unique identifier of feature type for the waterbody

lake_reachcode	reach code assigned to the water feature by the NHD

lake_namegnis	lake name from the gnis database

lake_namelagos	lake name from a combination of data sources (gnis, wqp, etc.)

lake_onlandborder	flag indicating lake polygon is adjacent to or crosses the border with Canada or Mexico

lake_ismultipart	flag indicating that the focal lake polygon is multipart

lake_missingws	flag indicating that the lake's watershed (both ws and nws) was not delineated

lake_shapeflag	flag indicating lake polygon shape is excessively angular (e.g., triangle, rectangle) or elongate (very thin relative to length); may indicate the lake is not natural in origin (angular) or is more riverine (elongate)

lake_lat_decdeg	latitude of centroid of the NHD lake polygon in decimal degrees (NAD83)	decimal degree

lake_lon_decdeg	longitude of centroid of the NHD lake polygon in decimal degrees (NAD83)	decimal degree

lake_elevation_m	the elevation of the lake polygon centroid, in meters (referenced to the North American Vertical Datum of 1988 (NAVD88) and obtained from the National Elevation Dataset)	meter

lake_centroidstate	abbreviation of state containing the lake centroid

lake_states	abbreviation(s) of state(s) intersecting the focal lake polygon

lake_county	NULLme of county containing the focal lake polygon centroid

lake_countyfips	Federal Information Processing Standard (FIPS) code for the county containing the focal lake polygon centroid

lake_huc12	the code for the HUC12 containing the lake centroid

buff100_zoneid	unique 100 m buffer identifier assigned by LAGOS-US, equivalent to the focal lake lagoslakeid

buff500_zoneid	unique 500 m buffer identifier assigned by LAGOS-US, equivalent to the focal lake lagoslakeid

ws_zoneid	unique watershed identifier assigned by LAGOS-US, equivalent to the focal lake lagoslakeid

nws_zoneid	unique network watershed identifier assigned by LAGOS-US, equivalent to the focal lake lagoslakeid

hu12_zoneid	unique identifier assigned by LAGOS-US for zones in the spatial division HU12

hu8_zoneid	unique identifier assigned by LAGOS-US for zones in the spatial division HU8

hu4_zoneid	unique identifier assigned by LAGOS-US for zones in the spatial division HU4

county_zoneid	unique identifier assigned by LAGOS-US for zones in the spatial division county

state_zoneid	unique identifier assigned by LAGOS-US for zones in the spatial division state

epanutr_zoneid	unique identifier assigned by LAGOS-US for zones in the spatial division EPA Nutrient Ecoregions

omernik3_zoneid	unique identifier assigned by LAGOS-US for zones in the spatial division Omernik Level III Ecoregions

wwf_zoneid	unique identifier assigned by LAGOS-US for zones in the spatial division World Wildlife Fund ecoregions

mlra_zoneid	unique identifier assigned by LAGOS-US for zones in the spatial division Major Land Areas

bailey_zoneid	unique identifier assigned by LAGOS-US for zones in the spatial division Bailey’s Ecoregions

neon_zoneid	unique identifier assigned by LAGOS-US for zones in the spatial division National Ecological Observation Network